Replace copy with symlink & little cleanup
So far we have a mess everywhere, but we are mainly using/working on
osc-staging.py file, so letting test suite use it by symlinking it for now.
General idea is once we cleanup everything, we get rid of all the mess around
and move this file to the correct location. Also moving some directories around
to clean up stuff even more.