Better testing, so far only for request reviews
To make testing more reliable and nice and generally better, we need to address
few issues. First of all, currently it is hard to figure out what was test
trying to access if accessing it fails. Second issue was that we mixed both
tests and OBS simulation logic. In perfect world, we should write some state
machine simulating OBS in some limited way and checking that scripts behaviour
corresponds to it.

This commit attempts to solve it at one example - review handling test.