staging accept: introduce --force
Allow to skip full validation of /totest being fully synced.
This can be helpful (and is only meant to be used) when you run two
consequtive staging accept LETTER commands in a row.

* The first one would see /totest fully sync, not dirty
* the second call sees /totest as dirty, as the scheduler would first have
  to process all data

In order to not have to wait for this to happen, the factory-maintainer CAN
override this with osc staging accept --force