repochecker: Adjust the decline message for missing binaries
Some users are confused as the old message "package never built" does
not have to be true in many cases, especially on longer-lasting
stagings. In this case, the binaries might have been built but have
since been replaced by newer ones - the effect is the same: repo-checker
can't receive them in order to perform the full checks.