Ugly workaround for glibc.i686 package config
Workaround glibc.i686 package config, it had a topadd block in _link,
and looks like it causes the disturl won't consistently with glibc even
with the same srcmd5. And the lastsuccess state will be outdated if the
revision is used the same srcmd5.

Note: the proper workaround, to instead of
5daf2bb6509abaf966268c9d6b13aa4e330a8af3