stagingapi: refactor supersede logic and handle different source projects.
Primarily an issue for Leap in that requests originating from different
source projects may be submitted for the same package at once. Instead of
always superseding the previous request the request will only be
superseded if it originates from the same project.

If different sources md5s are found the request will be ignored and left
for manual review and an identical request will be declined.