AcceptCommand: update version of all products found in the project
in opeSUSE:Factory, for example, there is also openSUS-Kubic.product and the
versions are supposed to stay in sync.

We newly find all product definitions inside a project and update all of them,
with the same limiation as before: only if version was a 8 digit version it will
be updated to the current date stamp

Additionally, this is a preparation for the new product builder, as _product is no
longer hardcoded, but found.