adi_command: do not checking adi staging status if there is no selected_requests
Wihtout selected_requests then that adi staging project is an empty project,
adi command should clean it up rather than go through the logic to check the
project status. Since we have introduced a kind of frozenlinks adi staging,
the project status might be build broken or so but it doesn't have any request
staged in that staging actually, those kind of projects must be deleted.