Fix rebuildpacs for armv6/armv7 architectures
For some architectures, the rpm architecture and the
Open Build Service Scheduler architecture is not identical. This
is partially unavoidable as they are defined differently (OBS
uses the kernel arch, and the rpm architecture might contain
extra ABI information that is only relevant for user space). Accomodate
by special casing the currently known two exceptions.