Support unneeded.yml as list of regular expressions
UNWANTED is a bit heavy weapon to filter unsorted.yml, so leave
its role to packages we sure don't want (e.g. -mini and -bootstrap
packages). Packages that are there but would end in unsorted otherwise,
filter by means of package regular expressions in unneeded.yml

The format of that file is yml with a list of groups of regular
expressions.

global:
  - .*-32bit

texlive:
  - texlive.*