Handle Leap:15.0:Ports images_arm repository
For Leap 15.0 Ports we need to be able to support two image
repos, one for arm and one for ppc because they get published
to two different locations. for now as we have only images_arm
add support for that while ppc is still making its way through
an initial build.