[oqamaint] Fix re-reviewing requests
We no longer need to be careful with requests that have started
jobs or not - we schedule the incident projects as the very first
action. So what's left to do is reviewing the requests and checking
if the current target repo jobs have the incident listed in some _TEST_ISSUES
variable. Otherwise the request is not to be reviewed