metrics/access/aggregate: write zeros for empty product data.
Since the aggregation is not done by influxdb a fill option is not
available and instead needs to be performed manually. As such the order of
writing must be from oldest to newest so that the list of product seen
before can be kept and missing products for a given interval can be filled.