leaper: remove only_one_action limitation.
Leaper will now be used on maintenance incidents which commonly have
multiple actions. A similar change was done to check_source.py where the
restriction was moved to a review setting which is more appropriate rather
than a hard limitation. Going forward it makes sense to lift the limitation
and have only one bot enforce the rule where appropriate.