metrics: remove workaround for OBS xpath bug during request search.
Now that the bug related to openSUSE/open-build-service#5571 has been fixed
the inverted behavior of exclude_target_projects no longer works. As such
the argument should be removed and only way to get the desired behavior
is to override include_request_from_project which includes request sourced
from specific project. For the purposes of metrics only interested in
requests targeting a specific project.