[ARM] Skip build number comparison
The LiveCD products (JeOS) for ARM will never follow
the same iso build numbers (and don't produce iso's to begin with)
than the main project. In lack of a "ignore mismatch for livecd
products only" option we just skip the syncing globally. that should
be fine as hopefully the main products always produce the same build.