dist/ci/Dockerfile: change from Tumbleweed to Leap for build host.
There is no need for Tumbleweed base as only the latest stable osc tools
and openSUSE:Tools packages should be needed. Unlike run-time for
repo-checker which is limited by latest solver stack the distribution
test suite has no need.

Currently, there is a zypper related segfault, #1887, that breaks the
test suite and demonstrates an advantage to switching to Leap.