Fix support for multiple spec files
Right after commit of a package into a staging project, it can happen
that OBS replies to GET with 'service in progress' and then we don't
get a file list. As we ignored that exception, we did not create the
other spec files.

Now enhance retried_GET to look for this condition and retry this as
well, so we get reliable package list.