StagingAPI: complex delete requests are no longer used
* Simplify the 'accept' code

openSUSE:Factory/snapshot is newly setup as a Download on Demand repository
and we can thus safe ourselves the hassle of having to take extra care of
handling delete requests. They no longer bleed directly into the /snapshot
area and can now be accepted like any other request

* Drop vdelreq plugin

Since we no longer do a virtual accept of delete requests, we also don't
have to monitor their state anymore. Drop the plugin that served only
this usecase

* Drop the acheck command

As a consequence, the command 'staging acheck' serves no purpose anymore:
/totest was just there to ensure we have something to copy to /snapshot.
This entire repo will vanish, and it is thus no longer needed to verify
if it is in sync before accepting new data.