Fix conflicts/obsoletes while picking dependencies
We ignored conflicts/obsoletes during pool preparation as it got
into the way of finding supplements (due to conflicting packages in
the list). But this way, conflicts were invisible also during dependency
search, leading to wrong packages picked for the package lists in
general.

So add a flag to prepare_pool to explain if we want conflicts or not