osclib/origin: origin_history(): include non-annotated requests.
Otherwise, requests that are declined quickly (likely by bot) are not
shown in history which can be confusing when viewing in web interface. A
package may show as being one revision behind and one would wonder why
not update request when in fact a request could have been made and
declined.