Staging accept: No longer set the version in .product files on accept
In the past, we had a dashboard reading the version of the current source from the .product file.
Not immediately updating this version after a checkin would have been confusing as the dashboard would
have shown incorrect/outdated information.

the new dashboard (https://factory-dashboard.opensuse.org) reads the source version directly from the
attribute, so this mis-match is eliminated.

The .product files will be updated by pkglistgen in due time.