Fix check_source for repository specific package names
If the package name as reported by OBS does not match the one
we're expecting, then loop through all repositories and check
if we find one there.

This is weakening the policy a little as this will open the
door for false negatives - e.g. that got the right package name
only for another repository. But as we do submission between
code streams all the time, I can't limit the package parsing
to repositories building against the target. So the opened
hole is to be closed by sanity check on review-team - as a
matter of fact the policy is not to catch people playing
macro games around Name, but for people that use completely
different names in source and target.

Fixes #2274