Fix staging-installcheck.py being ignored if i586 is present
For openSUSE:Factory:Staging:adi:*, i586 is only enabled when necessary and so
the check is not set as required. Currently x86_64 is handled first (due to
target_archs), which means that for oS:F, i586 comes last and so the ugly hack
results in x86_64 getting a fake success and the actual status is set for i586
only, which gets ignored.