legal: Only wait 2 hours for legal reviews for openSUSE:Factory
The churn rate of openSUSE:Factory is too high for legal to catch up
these days, so we turn the default. Legal reviews for Factory are no longer
blocking, but if we need the same package for another project or
product, the lawyers had some time to look at it.

And if we find a package not acceptable, we'll delete it from Factory
later on. This may get painful, but the current legal situation is
painful too.