pkglistgen: Delete sources after creating the final solv files
When a project is switched from `refresh: true` to `refresh: false`, the
`15.1:update_123XXX.packages.xz` files are kept alongside the new "final"
`15.1:update.packages.xz`. However, the `update_and_solve` command looks at all
of those, which there can be a lot of, and takes a long time to look at
effectively redundant data.

With this commit, the old files will be deleted for projects with
`refresh: false` once the final solv file is created. This is also done for
already "finalized" projects in a separate block.