Remove sentry integration
It was introduced by #2159 because the various machines were hard to
monitor. Since we moved to botmaster there is no longer a problem with
that and if it ever served us is hard to say as the sentry project's
password is lost for years. So simplify the exception handling by just
throwing it