Remove the origin operator and its user scripts
The service is down for 2 years and more and it was unused ever since.
The idea just never took off (also because it was pretty late in the
cycle and leap wasn't developed much longer the way it used to when
the design for this was created).

For the next SLE/Leap it's just very, very unclear how it will look
like and I'm 99% sure it's not going to go back to old Leap. So the
whole operator needs more work and maintenance than the current team
has, so keep it in the git history only.