Make frozenage less reliant on changing _frozenlinks
The mtime of a file in OBS is the time this content was created for the
first time. As such if the target project isn't moving for a week,
staging select will keep asking you to freeze - even if you just froze.

So instead of relying on OBS to tell the true mtime, set an attribute
in the staging project and read from it. This also gives a way to check
in the webui for the last freeze time

Fixes #2462