pkglistgen: Only solve one project at a time
As we deprecated 'all' we can just as well rely on only one project
given (which is also what we do in gocd) to ease error reporting.

The old method of running all projects in a loop stemed from the
old pkglistgen not having a scheduler so we relied on serial execution.