Survive broken requests
Revision openSUSE:Factory bash 49 obviously accepted Base:System rev 18
(revision 19 is "checked in"), but the request associated with revision 49
(26846) does not bear that information *and* got invalid timestamp of 2016
(which this is happened in 2009).

So we need to be creative - and the way I came up with is returning HEAD
of the devel branch for such requests.