skippkg-finder: add conditionals support for the product ftp-tree
In some cases we would not like to add package to ftp-tree on an
specific arch, this can be done by multiple 'group' to be defined in
NON_FTP_PACKAGES.group, however NON_FTP_PACKAGES.group is auto-generated
by skippkg-finder according the buildresults and the project attributes,
this commit will add conditionals support to skippkg-finder, in other
words, to be able to define arch dependent package in
NON_FTP_PACKAGES.group via the project attribute.

To enable it, 'use group' is also needed to be existsed in the product
file, the available condition/platform can be found in meta pacakge.