Fix the OBS CI workflow (#2950)
The package needs to have the obs_scm service enabled to fetch the right sources.
For submissions to oS:F this is not possible so the devel pkg can't be used.
Use a package meant for this instead. This way a branch can be used as well,
so the project setup can be taken from the prj on OBS.