Add BCI repo publishing bot
Unfortunately TTM itself doesn't quite fit here due to the reasons outlined
in the code and the additional checks are not a good fit for adding it into
the go pipeline description, so write yet another OBS publishing bot!

For the initial test runs this is not triggered by timer runs yet.