Attempt to publish BCI repos more often
The goal is to publish the repo at most daily. Previously this was ensured
by simply running the bot only once a day. However, due to OBS and openQA
timing this was frequently attempted when either of those wasn't done yet,
which happened several times in a row. To not miss times when they align,
attempt publishing more often but skip a build if the previous publish was
recent enough.