tests: checktags_tests: support multiple factories
Multiple factories can be specified in ALLOWED_FACTORIES. Tests are now
supposed to specify which factory they want to test on via the factories
keyword in _run_with_data(), exists_in_factory has been removed.

The data is identical. If no factory is specified, it behaves the same
way as with exists_in_factory=False.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>