publish bci repo right away on package set changes
When there was a source commit on the bci repo, then the set of
packages changed, which likely means we need those changes to pass
the tests on the container release pipelines. So don't wait for 24h
but publish right away.

Also trigger on the oldest published state rather than the newest one
so that we do not need to wait another 24h per openQA failure to resume
publishing.