gocd: sp/micro-targets: ensure the pipeline exits on failures
This was already implemented in ddd2304022e313eca9b391a043b12ecb9afe3bf7
for SLE GA to TEST, but missing in Micro and on the TEST -> PUBLISH leg
in SLE.

Releasing from TEST to PUBLISH might still fail for various reasons.
If it happens, ensure the pipeline exits early with non-zero exit code.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>