slfo-stagings: add a timeout for Build.product, and do not run if repo is missing
Ensure that the 'product' repo is configured in the project meta.

Unfortunately, since we're checking in the configuration of the staging
itself, rather than the parent project's, there are increased changes of
some human/tooling error that might drop the repository from the
configuration. This will make the job run an endless loop.

Also, add a timeout for 180 minutes for the job. It is enough for
the product build process, and if it isn't, it's better to cancel
and have human eyes looking at it.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>