tests: OBSLocal: do not pick up __name__ when updating existing config
The __name__ option is added by OscConfigParser automatically, and
it contains the section name.

When updating the configuration in FactoryWorkflow.remote_config_set(),
__name__ would be fetched as part of the existing configuration that
needs to be updated and would in turn be submitted again.

Since osc 1.8.0, having duplicate options is not supported [0].
Let's skip __name__ when updating the mock configuration, as it's
an internal option that we don't want to touch anyway.

[0] https://github.com/openSUSE/osc/commit/360a94c4a3745dfe59863c6a112acf8bcf38840d

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>