legal-auto: Exclude openSUSE:Factory from commenting
Since there have been quite a few vocal complaints about comments
with links to LegalDB reviews not being useful for
openSUSE:Factory, we now exclude the project.

I've also rephrased review messages in a consistent format, and
emphasised the fact that the opensuse-review-team is responsible for
license compliance if the 2 hour timeout is reached without a legal
review.