Look for images and sbom assets also in the iso/ folder
After the latest changes of product-composer (allowing to build isos with different
names than ftp trees), the sbom assets are no longer saved in the root of the repo
(as they are for SLFO 1.0) and now live in the iso folder as well.

So look for those assets in both places.