packaging: Do not reference python3-osc
This is only a symbol provided by the actual osc package.

One of the main issues attempted to be addressed was ensuring that the
plugin matches the python version used by osc, but the symbol
'python3-*' does not guarantee that at all, as the symbol just moves between
python-flavors, based on primary python.