osclib: core: explicitly replace OBS instance prefix in devel_project_fallback
When getting the devel project via the build service API, and the source
project is accessed via interconnect, the returned devel project
contains the instance prefix.

However, when checking directly in the Factory git mapping, the
interconnected prefix is not being added to the result.

As such, ensure we handle this case too - and just replace the instance
name if found.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>