fix: quote variables to prevent wildcard expansion
When a development project cannot be determined, the script sets the project
variable to "***** UNKNOWN". Previously, this variable was used unquoted in an
echo command, causing the shell to expand "*****" into a list of all files in
the current directory. This resulted in corrupted entries in the devel_packages
file e.g. as follows:
```
lightdm-kde-greeter cat devel_packages devel_packages.254071 echo gpg-agent killall UNKNOWN
```